refactor: consolidate task module layout
