+++ src-tauri/src/lib.rs
@@ -12,10 +12,10 @@ pub mod portable;
 pub mod resolution;
 pub(crate) mod scan_runtime;
 pub mod scene;
+pub mod screenshot;
 pub mod session;
 pub mod sound;
 pub mod task;
-pub mod tasks;
 pub mod tauri_commands;
 pub mod template_matching;
 pub mod tray;
