+++ src-tauri/src/session/mod.rs
@@ -6,7 +6,7 @@
 //!
 //! 职责边界：
 //! - ✅ 只做薄委托与坐标缩放，**不含任何业务逻辑**（不识别场景、不导航、不扫描）；
-//! - ✅ **不依赖前端**（结果上报由任务层通过 [`crate::tasks::archive_scan::ScanReporter`] 完成）；
+//! - ✅ **不依赖前端**（结果上报由任务层通过 [`crate::task::archive_scan::ScanReporter`] 完成）；
 //! - ✅ 只依赖基础设施层。
 //!
 //! 会话贯穿一次游戏操作（扫描档案库任务），由调用方以 `&mut` 串行使用。
