+++ src-tauri/src/task/archive_scan/mod.rs
