+++ src-tauri/src/task/archive_scan/plan.rs
