+++ src-tauri/src/task/archive_scan/result.rs
