+++ src-tauri/src/task/archive_scan/scan_loop.rs
