+++ src-tauri/src/task/archive_scan/task.rs
