--- src-tauri/src/tasks/mod.rs
@@ -1,4 +1,0 @@
-//! 任务模块：存放所有具体的游戏自动化任务实现。
-
-pub mod archive_scan;
-pub mod screenshot;
